Made own seeds file
