+++ Gemfile
@@ -1,0 +1,5 @@
+source "https://rubygems.org"
+
+gem "pg"  # this gem allows ruby to talk to postgres
+gem "activerecord"  # this gem provides a connection between your ruby classes to relational database tables
+gem "pry"  # this gem allows access to REPL
