+++ db/connection.rb
@@ -1,0 +1,8 @@
+require "pg" # postgres db library
+require "active_record" # the ORM
+require "pry" # for debugging
+
+ActiveRecord::Base.establish_connection(
+  :adapter => "postgresql",
+  :database => "landlord"
+)
