+++ models/apartment.rb
@@ -1,0 +1,3 @@
+class Apartment < ActiveRecord::Base
+  has_many :tenants
+end
