+++ models/tenant.rb
@@ -1,0 +1,3 @@
+class Tenant < ActiveRecord::Base
+  belongs_to :apartment
+end
